fix: address review feedback for #68
