+++ src/client/lib/hooks/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Shared Hooks
+ *
+ * Re-exports all shared hooks for easy importing.
+ */
+
+export { useIOSKeyboardOffset } from './useIOSKeyboardOffset';
