+++ src/client/routes/Profile/components/EditableField.tsx
@@ -9,44 +9,7 @@ import { Input } from '@/client/components/ui/input';
 import { Button } from '@/client/components/ui/button';
 import { Sheet, SheetContent, SheetHeader, SheetTitle } from '@/client/components/ui/sheet';
 import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogTrigger } from '@/client/components/ui/dialog';
-
-/**
- * Hook to detect iOS keyboard and calculate offset needed to keep content visible.
- * Uses visualViewport API to detect when keyboard reduces visible area.
- */
-function useIOSKeyboardOffset() {
-    // eslint-disable-next-line state-management/prefer-state-architecture -- ephemeral keyboard state
-    const [keyboardOffset, setKeyboardOffset] = useState(0);
-
-    useEffect(() => {
-        // Only run in browser and when visualViewport is available
-        if (typeof window === 'undefined' || !window.visualViewport) {
-            return;
-        }
-
-        const viewport = window.visualViewport;
-
-        const handleResize = () => {
-            // Calculate the difference between window height and viewport height
-            // This difference is the keyboard height on iOS
-            const offset = window.innerHeight - viewport.height;
-            setKeyboardOffset(offset > 0 ? offset : 0);
-        };
-
-        viewport.addEventListener('resize', handleResize);
-        viewport.addEventListener('scroll', handleResize);
-
-        // Initial check
-        handleResize();
-
-        return () => {
-            viewport.removeEventListener('resize', handleResize);
-            viewport.removeEventListener('scroll', handleResize);
-        };
-    }, []);
-
-    return keyboardOffset;
-}
+import { useIOSKeyboardOffset } from '@/client/lib/hooks';
 
 interface EditableFieldProps {
     label: string;
